# Remove use of notty for formatting benchmark results