--- a/bench/bechamel_csv/bechamel_csv.mli
+++ b/bench/bechamel_csv/bechamel_csv.mli
@@ -1,0 +1,2 @@
+val pp : (string, (string, Bechamel.Analyze.OLS.t) Stdlib.Hashtbl.t) Stdlib.Hashtbl.t Fmt.t
+(** [pp] formats the results of {!Bechamel.Analyze.merge} as a CSV table. *)