--- a/bench/bechamel_csv/dune
+++ b/bench/bechamel_csv/dune
@@ -1,0 +1,3 @@
+(library
+  (name bechamel_csv)
+  (libraries bechamel fmt))