--- a/bench/cptest.ml
+++ b/bench/cptest.ml
@@ -73,20 +73,5 @@ let benchmark () =
       (results, raw_results)
 
 let () =
-  List.iter
-    (fun v -> Bechamel_notty.Unit.add v (Measure.unit v))
-    Instance.[ minor_allocated; major_allocated; monotonic_clock ]
-
-let img (window, results) =
-  Bechamel_notty.Multiple.image_of_ols_results ~rect:window
-    ~predictor:Measure.run results
-
-open Notty_unix
-    
-let () =
-  let window =
-    match winsize Unix.stdout with
-    | Some (w, h) -> { Bechamel_notty.w; h }
-    | None -> { Bechamel_notty.w = 80; h = 1 } in
   let results, _ = benchmark () in
-  img (window, results) |> eol |> output_image
+  Fmt.pr "%a@." Bechamel_csv.pp results