--- a/dune-project
+++ b/dune-project
@@ -17,8 +17,6 @@
   (ocaml (>= 4.12.0))
   dune-configurator
   (lwt (and :with-test (>= 5.0.0)))
-  (notty (and (>= 0.2.2) :with-test))
-  (bechamel-notty (and (>= 0.1.0) :with-test))
   (bechamel (and (>= 0.1.0) :with-test))
   (logs (and :with-test (>= 0.5.0)))
   (cmdliner (and :with-test (>= 1.1.0)))