--- a/uring.opam
+++ b/uring.opam
@@ -15,8 +15,6 @@ depends: [
   "ocaml" {>= "4.12.0"}
   "dune-configurator"
   "lwt" {with-test & >= "5.0.0"}
-  "notty" {>= "0.2.2" & with-test}
-  "bechamel-notty" {>= "0.1.0" & with-test}
   "bechamel" {>= "0.1.0" & with-test}
   "logs" {with-test & >= "0.5.0"}
   "cmdliner" {with-test & >= "1.1.0"}