The dependency on notty prevents uring from being tested on OCaml 5.

Old output:
```
╭────────────────┬───────────────────────────┬───────────────────────────┬───────────────────────────╮
│name            │  major-allocated          │  minor-allocated          │  monotonic-clock          │
├────────────────┼───────────────────────────┼───────────────────────────┼───────────────────────────┤
│  noop      10  │             0.0000 mjw/run│           160.2533 mnw/run│           1003.7415 ns/run│
│  noop      30  │             0.0000 mjw/run│           480.3344 mnw/run│           2687.7522 ns/run│
│  noop     100  │             0.0930 mjw/run│          1600.4478 mnw/run│           9072.9584 ns/run│
│  noop     300  │            12.2791 mjw/run│          4800.6122 mnw/run│          27520.1296 ns/run│
│  noop    1000  │           114.9145 mjw/run│         16000.9554 mnw/run│          98416.3581 ns/run│
│  noop    3000  │          1030.7835 mjw/run│         48001.5152 mnw/run│         298576.3733 ns/run│
│  noop   10000  │         13413.3058 mjw/run│        160002.6316 mnw/run│        1078551.5092 ns/run│
╰────────────────┴───────────────────────────┴───────────────────────────┴───────────────────────────╯
```
New output:
```
          name,   major-allocated,   minor-allocated,   monotonic-clock
  noop      10,          0.003163,         89.035794,       1122.076257
  noop      30,          0.010071,        272.687218,       2796.151085
  noop     100,          0.914579,        978.710714,       9079.485427
  noop     300,         11.478316,       3511.237206,      28500.701765
  noop    1000,        115.853532,      13936.387169,     104048.680888
  noop    3000,       1033.692225,      44261.436753,     308647.065684
  noop   10000,      12420.745528,     153296.414960,    1216105.351957
```
The units have gone, but on the plus side you can now graph it as CSV data.

The results do seem to be slightly different, though, which is odd. Hmm. Restoring the `List.iter (fun v -> Bechamel_notty.Unit.add v (Measure.unit v)) metrics;` line affects things:

```
          name,   major-allocated,   minor-allocated,   monotonic-clock
  noop      10,          0.000000,        160.284696,       1124.221385
  noop      30,          0.000000,        480.360509,       2997.389849
  noop     100,          0.000000,       1600.467290,       9834.411904
  noop     300,         12.125976,       4800.649351,      30976.050309
  noop    1000,        112.719320,      16001.034483,     115284.222115
  noop    3000,       1022.773418,      48001.612903,     339099.821312
  noop   10000,      13348.487663,     160002.830189,    1201494.350105
```

@dinosaure: is this the right way to do this?