@talex5 2022-08-24
A couple of the new test platforms fail with:
```
 | /tmp/build_200d78_dune/ocaml-configuratorf12775/c-test-0/test.c:5:10: fatal error: linux/time_types.h: No such file or directory
 |  #include <linux/time_types.h>
 |           ^~~~~~~~~~~~~~~~~~~~
 | compilation terminated.
Error: failed to compile program
```
I think we can remove this once #70 is merged, since liburing.h pulls it in correctly (cc @bikallem).

Also, it appears that #37 is fixed now!