@bikallem 2022-08-24
> I think we can remove this once https://github.com/ocaml-multicore/ocaml-uring/pull/70 is merged, since liburing.h pulls it in correctly

Yes, "liburing.h" defines `__kernel_timespec` if it is not found. https://github.com/axboe/liburing/blob/master/configure#L446-L453

I originally intended to just use "liburing.h" with dune configurator but wasn't successful. So #70 should make "<linux/time_types.h>" redundant and can be removed.